+++ sdk/core/core-rest-pipeline/src/index.ts
@@ -20,6 +20,7 @@ export {
   RawHttpHeadersInput,
   RequestBodyType,
   SendRequest,
+  StreamBuilder,
   TlsSettings,
   TransferProgressEvent,
 } from "./interfaces";