+++ sdk/core/core-rest-pipeline/src/policies/multipartPolicy.ts
@@ -2,7 +2,13 @@
 // Licensed under the MIT license.
 
 import { randomUUID, stringToUint8Array } from "@azure/core-util";
-import { BodyPart, HttpHeaders, PipelineRequest, RequestBodyType } from "../interfaces";
+import {
+  BodyPart,
+  HttpHeaders,
+  PipelineRequest,
+  RequestBodyType,
+  StreamBuilder,
+} from "../interfaces";
 import { PipelinePolicy } from "../pipeline";
 import { toStream, concatenateStreams } from "../util/stream";
 import { isBlob } from "../util/typeGuards";
@@ -20,13 +26,7 @@ function encodeHeaders(headers: HttpHeaders): string {
 }
 
 function getLength(
-  source:
-    | (() => ReadableStream<Uint8Array>)
-    | (() => NodeJS.ReadableStream)
-    | Uint8Array
-    | Blob
-    | ReadableStream
-    | NodeJS.ReadableStream
+  source: StreamBuilder | Uint8Array | Blob | ReadableStream | NodeJS.ReadableStream
 ): number | undefined {
   if (source instanceof Uint8Array) {
     return source.byteLength;
@@ -39,14 +39,7 @@ function getLength(
 }
 
 function getTotalLength(
-  sources: (
-    | (() => ReadableStream<Uint8Array>)
-    | (() => NodeJS.ReadableStream)
-    | Uint8Array
-    | Blob
-    | ReadableStream
-    | NodeJS.ReadableStream
-  )[]
+  sources: (StreamBuilder | Uint8Array | Blob | ReadableStream | NodeJS.ReadableStream)[]
 ): number | undefined {
   let total = 0;
   for (const source of sources) {