+++ sdk/openai/openai/src/api/operations.ts
@@ -55,7 +55,7 @@ import {
   GetAudioTranslationOptions,
 } from "../models/audio.js";
 import { renameKeysToCamelCase } from "./util.js";
-import { createFile } from "@azure/core-rest-pipeline";
+import { StreamBuilder, createFileFromStream } from "@azure/core-rest-pipeline";
 
 export function _getEmbeddingsSend(
   context: Client,
@@ -665,21 +665,21 @@ function _getChatCompletionsSendX(
  * Returns the translation of an audio file.
  * @param context - The context containing the client to use for this request.
  * @param deploymentName - The name of the model deployment (when using Azure OpenAI) or model name (when using non-Azure OpenAI) to use for this request.
- * @param fileContent - The content of the audio file to translate.
+ * @param buildFileStream - A function to return a stream of the audio content to translate.
  * @param options - The options for this audio translation request.
  * @returns The audio translation result.
  */
 export async function getAudioTranslation(
   context: Client,
   deploymentName: string,
-  fileContent: Uint8Array,
+  buildFileStream: StreamBuilder,
   options?: GetAudioTranslationOptions
 ): Promise<AudioResultSimpleJson>;
 /**
  * Returns the translation of an audio file.
  * @param context - The context containing the client to use for this request.
  * @param deploymentName - The name of the model deployment (when using Azure OpenAI) or model name (when using non-Azure OpenAI) to use for this request.
- * @param fileContent - The content of the audio file to translate.
+ * @param buildFileStream - A function to return a stream of the audio content to translate.
  * @param format - The format of the result object. See {@link AudioResultFormat} for possible values.
  * @param options - The options for this audio translation request.
  * @returns The audio translation result.
@@ -687,7 +687,7 @@ export async function getAudioTranslation(
 export async function getAudioTranslation<Format extends AudioResultFormat>(
   context: Client,
   deploymentName: string,
-  fileContent: Uint8Array,
+  buildFileStream: StreamBuilder,
   format: Format,
   options?: GetAudioTranslationOptions
 ): Promise<AudioResult<Format>>;
@@ -695,7 +695,7 @@ export async function getAudioTranslation<Format extends AudioResultFormat>(
 export async function getAudioTranslation<Format extends AudioResultFormat>(
   context: Client,
   deploymentName: string,
-  fileContent: Uint8Array,
+  buildFileStream: StreamBuilder,
   formatOrOptions?: Format | GetAudioTranslationOptions,
   inputOptions?: GetAudioTranslationOptions
 ): Promise<AudioResult<Format>> {
@@ -707,7 +707,7 @@ export async function getAudioTranslation<Format extends AudioResultFormat>(
     .pathUnchecked("deployments/{deploymentId}/audio/translations", deploymentName)
     .post({
       body: {
-        file: createFile(fileContent, "placeholder.wav"),
+        file: createFileFromStream(buildFileStream, "placeholder.wav"),
         ...(response_format && { response_format }),
         ...(temperature !== undefined ? { temperature } : {}),
         ...(prompt && { prompt }),
@@ -728,21 +728,21 @@ export async function getAudioTranslation<Format extends AudioResultFormat>(
  * Returns the transcription of an audio file in a simple JSON format.
  * @param context - The context containing the client to use for this request.
  * @param deploymentName - The name of the model deployment (when using Azure OpenAI) or model name (when using non-Azure OpenAI) to use for this request.
- * @param fileContent - The content of the audio file to transcribe.
+ * @param buildFileStream - A function to return a stream of the audio content to transcribe.
  * @param options - The options for this audio transcription request.
  * @returns The audio transcription result in a simple JSON format.
  */
 export async function getAudioTranscription(
   context: Client,
   deploymentName: string,
-  fileContent: Uint8Array,
+  buildFileStream: StreamBuilder,
   options?: GetAudioTranscriptionOptions
 ): Promise<AudioResultSimpleJson>;
 /**
  * Returns the transcription of an audio file.
  * @param context - The context containing the client to use for this request.
  * @param deploymentName - The name of the model deployment (when using Azure OpenAI) or model name (when using non-Azure OpenAI) to use for this request.
- * @param fileContent - The content of the audio file to transcribe.
+ * @param buildFileStream - A function to return a stream of the audio content to transcribe.
  * @param format - The format of the result object. See {@link AudioResultFormat} for possible values.
  * @param options - The options for this audio transcription request.
  * @returns The audio transcription result in a format of your choice.
@@ -750,7 +750,7 @@ export async function getAudioTranscription(
 export async function getAudioTranscription<Format extends AudioResultFormat>(
   context: Client,
   deploymentName: string,
-  fileContent: Uint8Array,
+  buildFileStream: StreamBuilder,
   format: Format,
   options?: GetAudioTranscriptionOptions
 ): Promise<AudioResult<Format>>;
@@ -758,7 +758,7 @@ export async function getAudioTranscription<Format extends AudioResultFormat>(
 export async function getAudioTranscription<Format extends AudioResultFormat>(
   context: Client,
   deploymentName: string,
-  fileContent: Uint8Array,
+  buildFileStream: StreamBuilder,
   formatOrOptions?: Format | GetAudioTranscriptionOptions,
   inputOptions?: GetAudioTranscriptionOptions
 ): Promise<AudioResult<Format>> {
@@ -770,7 +770,7 @@ export async function getAudioTranscription<Format extends AudioResultFormat>(
     .pathUnchecked("deployments/{deploymentId}/audio/transcriptions", deploymentName)
     .post({
       body: {
-        file: createFile(fileContent, "placeholder.wav"),
+        file: createFileFromStream(buildFileStream, "placeholder.wav"),
         ...(response_format && { response_format }),
         ...(language && { language }),
         ...(temperature !== undefined ? { temperature } : {}),